# app/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Fisher Fans API
  description: API pour gérer les utilisateurs, bateaux, sorties de pêche, réservations et carnets de pêche.
  version: 1.0.0
servers:
  - url: https://api.fisherfans.com/
    description: Serveur de production Fisher Fans
  - url: https://localhost:3000/
    description: Serveur de développement Fisher Fans
paths:
...
  /v1/reservations:
    post:
      summary: Créer une nouvelle réservation
      tags:
        - Reservations
      requestBody:
        $ref: '#/components/requestBodies/Reservation'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /v1/reservations/{id}:
    get:
      summary: Obtenir les détails d'une réservation
      tags:
        - Reservations
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      responses:
        '200':
          $ref: '#/components/schemas/Reservation'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    put:
      summary: Modifier une réservation
      tags:
        - Reservations
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      requestBody:
        $ref: '#/components/requestBodies/Reservation'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      summary: Annuler une réservation
      tags:
        - Reservations
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  responses:
    SuccessResponse:
      description: Opération réussie
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    ErrorResponse:
      description: Requête incorrecte ou ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  requestBodies:
...
    Reservation:
      description: Détails de la réservation
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Reservation'
...
  parameters:
...
    ReservationId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID unique de la réservation
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: ID unique de l'utilisateur
        name:
          type: string
          description: Nom de l'utilisateur
        email:
          type: string
          description: Adresse e-mail de l'utilisateur
        role:
          type: string
          description: Rôle de l'utilisateur (particulier, professionnel)
          enum: [particulier, professionnel]
        boats:
          type: array
          items:
            $ref: '#/components/schemas/Boat'
    Boat:
      type: object
      properties:
        id:
          type: integer
          description: ID unique du bateau
        name:
          type: string
          description: Nom du bateau
        boat_type:
          type: string
          description: Type du bateau
          enum: [open, cabin, catamaran, yacht, jetski, canoe]
        motor:
          type: string
          description: Type de motorisation
          enum: [diesel, essence, none]
        capacity:
          type: integer
          description: Capacité du bateau
        owner:
          $ref: '#/components/schemas/User'
    Trip:
      type: object
      properties:
        id:
          type: integer
          description: ID unique de la sortie de pêche
        title:
          type: string
          description: Titre de la sortie de pêche
        trip_date:
          type: string
          format: date
          description: Date de la sortie
        price:
          type: number
          format: float
          description: Prix de la sortie de pêche
        boat:
          $ref: '#/components/schemas/Boat'
        organizer:
          $ref: '#/components/schemas/User'
    Reservation:
      type: object
      properties:
        id:
          type: integer
          description: ID unique de la réservation
        user:
          $ref: '#/components/schemas/User'
        trip:
          $ref: '#/components/schemas/Trip'
        number_of_seats:
          type: integer
          description: Nombre de places réservées
        total_price:
          type: number
          format: float
          description: Prix total de la réservation
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicateur de succès de l'opération
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Message d'erreur
...